remove invalid tests and better readbility
