+++ nemo_automodel/components/distributed/pipelining/autopipeline.py
@@ -70,8 +70,6 @@ class AutoPipeline:
         patch_inner_model: bool = True,
         patch_causal_lm_model: bool = True,
         patch_stage_backward_maybe_with_nosync: bool = False,
-        # Tensor / Sequence Parallelism
-        sequence_parallel: bool = False,
         # Runtime
         device: Optional[torch.device] = None,
         dtype: Optional[torch.dtype] = None,
@@ -106,7 +104,6 @@ class AutoPipeline:
         self.patch_inner_model = patch_inner_model
         self.patch_causal_lm_model = patch_causal_lm_model
         self.patch_stage_backward_maybe_with_nosync = patch_stage_backward_maybe_with_nosync
-        self.sequence_parallel = sequence_parallel
         self._device: torch.device = device or torch.device("cuda" if torch.cuda.is_available() else "cpu")
         self.dtype = dtype
         self.scale_grads_in_schedule = scale_grads_in_schedule
@@ -164,7 +161,6 @@ class AutoPipeline:
             round_to_pp_multiple=self.round_virtual_stages_to_pp_multiple,
             patch_stage_backward_maybe_with_nosync=self.patch_stage_backward_maybe_with_nosync,
             seq_len=self.pp_seq_len,
-            sequence_parallel=self.sequence_parallel,
         )
 
         # Update PipelineInfo state
