+++ tests/functional_tests/hf_peft/test_hf_peft.py
@@ -23,7 +23,6 @@ HF_PEFT_VLM_FSDP2_CHECKPOINT_FILENAME = "L2_HF_PEFT_VLM_FSDP2_Checkpoint.sh"
 HF_QLORA_TINY_FILENAME = "L2_HF_QLORA_Tiny.sh"
 HF_PEFT_DEEPSEEK_MOE_LORA_FILENAME = "L2_HF_PEFT_DeepSeek_MoE_LoRA.sh"
 HF_PEFT_DORA_FSDP2_CHECKPOINT_FILENAME = "L2_HF_PEFT_DoRA_FSDP2_Checkpoint.sh"
-HF_PEFT_FUSED_QKV_CHECKPOINT_FILENAME = "L2_HF_PEFT_Fused_QKV_Checkpoint.sh"
 HF_MERGE_LORA_DENSE_FILENAME = "L2_Merge_LoRA_Dense.sh"
 HF_MERGE_LORA_MOE_FILENAME = "L2_Merge_LoRA_MoE.sh"
 
@@ -75,13 +74,6 @@ class TestHFPEFT:
         finally:
             # remove the checkpoint directory
             shutil.rmtree("checkpoints/", ignore_errors=True)
-
-    def test_hf_peft_fused_qkv_checkpoint(self):
-        try:
-            run_test_script(TEST_FOLDER, HF_PEFT_FUSED_QKV_CHECKPOINT_FILENAME)
-        finally:
-            # remove the checkpoint directory
-            shutil.rmtree("checkpoints_peft_fused_qkv_test/", ignore_errors=True)
 
     def test_merge_lora_dense(self):
         run_test_script(TEST_FOLDER, HF_MERGE_LORA_DENSE_FILENAME)
